tidied up code according to ms conventions and standards

+++ Portfolio_MauiNewsfeed/Services/GenericAppDataService.cs
@@ -10,8 +10,12 @@ namespace Portfolio_MauiNewsfeed.Services
     public class GenericAppDataService<T> : IAppDataService<T> where T : IAppDataItem
     {
         public string EntityExtension { get; set; }
+
         public virtual async Task<List<T>> GetAllAsync()
         {
+            if (EntityExtension == string.Empty)
+                throw new InvalidOperationException("Service requires an Entity Extension to be defined for getting files.");
+
             List<T> entities = new List<T>();
             IEnumerable<string> filterFilenames = Directory.GetFiles(FileSystem.AppDataDirectory).Where(x => x.EndsWith(EntityExtension));
 
@@ -26,6 +30,9 @@ namespace Portfolio_MauiNewsfeed.Services
 
         public virtual async Task SaveAsync(T entity)
         {
+            if (EntityExtension == string.Empty)
+                throw new InvalidOperationException("Service requires an Entity Extension to be defined for saving files.");
+
             string fileName = Path.Combine(FileSystem.AppDataDirectory, entity.Title + EntityExtension);
             var serializedData = JsonSerializer.Serialize(entity);
             await File.WriteAllTextAsync(fileName, serializedData);
@@ -33,9 +40,11 @@ namespace Portfolio_MauiNewsfeed.Services
 
         public virtual void Delete(T entity)
         {
+            if (EntityExtension == string.Empty)
+                throw new InvalidOperationException("Service requires an Entity Extension to be defined for deleting files.");
+
             string fileName = Path.Combine(FileSystem.AppDataDirectory, entity.Title + EntityExtension);
             File.Delete(fileName);
         }
-
     }
 }
